requirements.txt
#playwright==1.48.0
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
requests
firecrawl-py
